Updates the amount used in the payment initiation examples
Without this fix the first atttempt at using the quickstart in the development environment seems doomed to fail if the paymnet initiation limit is set below GBP12.34. This PR reduces the amount to hopefully reduce the likelihood of needing to contact support to track down the error that occurs if the amount it too high.